# speedup build time by using cotire which generates precompiled headers