Adds support for automatic compile time reduction provided by cotire. The feature can be deactivated using the flag `ENABLE_COMPTIME_RED`. Per default the flag is only enabled if using GNU compiler, as icc cannot use it with the specified compiler flags.

A short comparison using gcc shows a significant speedup.
~~~
Without cotire
make -j 8  251,35s user 17,63s system 669% cpu 40,151 total

With cotire
make -j 8  200,96s user 15,58s system 583% cpu 37,121 total
~~~